pull version assert into helper, fix meta issues
